integrate prototype
