+++ traversal/src/test/scala/overflowdb/traversal/RepeatTraversalTests.scala
@@ -238,132 +238,5 @@ class RepeatTraversalTests extends WordSpec with Matchers {
       //        __().out().asInstanceOf[org.apache.tinkerpop.gremlin.process.traversal.Traversal[_, Node]]
       //      ).times(repeatCount).values[String](Name.name).next() shouldBe "b"
     }
-
-    "foo2" in {
-      val repeatCount = 9000
-      //      val repeatCount = 900
-      val trav = Traversal.fromSingle(a).repeatDfs2(_.out, new RepeatBehaviour.Builder[Node].times(repeatCount).build)
-      //      val trav = Traversal.fromSingle(a).repeat8(_.out)(_.times(repeatCount))
-      println("first: " + trav.next)
-      trav.hasNext //stackoverflow
-    }
-
-    "foo3" in {
-      val repeatCount = 2
-      val trav = Traversal.fromSingle(a).repeatDfs3(_.out, new RepeatBehaviour.Builder[Node].times(repeatCount).build).property(Name)
-      println("hasNext: " + trav.hasNext)
-      //      println("first: " + trav.next)
-      //      println("hasNext: " + trav.hasNext)
-    }
-
-    "foo4" in {
-      centerTrav.foreach { center =>
-        println(center.property(Name))
-        center.out.foreach { one =>
-          println(one.property(Name))
-          one.out.foreach { two =>
-            println(two.property(Name))
-          }
-        }
-      }
-      // this is DFS: Center, L1, L2, R1, R2
-      // now make it lazy
-    }
-
-    "foo5 - this works!" in {
-      type Depth = Int
-      val stack: mutable.Stack[(Traversal[Thing], Depth)] = mutable.Stack.empty
-      stack.push((centerTrav, 0))
-
-      val untilDepth = 3
-
-      val iter = new Iterator[Thing] {
-
-        override def hasNext: Boolean = {
-          while (stack.nonEmpty) {
-            val (trav, depth) = stack.top
-            if (trav.isEmpty) stack.pop()
-            else if (depth == untilDepth) return true
-            else {
-              val element = trav.next
-              println(s"traversing over $element")
-              stack.push((Traversal.fromSingle(element).followedBy, depth + 1))
-            }
-          }
-          false
-        }
-
-        override def next(): Thing = {
-          if (hasNext) {
-            // at this point it's guaranteed to have the next non-empty traversal at the top of the stack
-            stack.top._1.next
-          } else {
-            throw new NoSuchElementException("next on empty iterator")
-          }
-        }
-      }
-
-      println(iter.next)
-      println(iter.next)
-      println(iter.hasNext)
-    }
-
-    "foo6 - try with circular graph" in {
-      // using circular graph so that we can repeat any number of times
-      val graph = SimpleDomain.newGraph
-
-      def addThing(name: String) = graph + (Thing.Label, Name -> name)
-
-      val a = addThing("a")
-      val b = addThing("b")
-      val c = addThing("c")
-      a --- Connection.Label --> b
-      b --- Connection.Label --> c
-      c --- Connection.Label --> a
-
-//      val repeatCount = 2
-//      val repeatCount = 20
-      val repeatCount = 100000
-//      Traversal.fromSingle(a).repeatX(_.out)(_.times(repeatCount)).property(Name).l shouldBe List("b")
-//        Traversal.fromSingle(a).repeatX(_.out)(_.times(repeatCount).breadthFirstSearch).property(Name).l shouldBe List("b")
-
-      type Depth = Int
-      val stack: mutable.Stack[(Traversal[Node], Depth)] = mutable.Stack.empty
-      val startTraversal = Traversal.fromSingle(a)
-      stack.push((startTraversal, 0))
-
-      val iter = new Iterator[Node] {
-
-        override def hasNext: Boolean = {
-          while (stack.nonEmpty) {
-            val (trav, depth) = stack.top
-            if (trav.isEmpty) stack.pop()
-            else if (depth == repeatCount) return true
-            else {
-              val element = trav.next
-              println(s"traversing over $element")
-              stack.push((Traversal.fromSingle(element).out, depth + 1))
-            }
-          }
-          false
-        }
-
-        override def next(): Node = {
-          if (hasNext) {
-            // at this point it's guaranteed to have the next non-empty traversal at the top of the stack
-            stack.top._1.next
-          } else {
-            throw new NoSuchElementException("next on empty iterator")
-          }
-        }
-      }
-
-
-      println(iter.toList)
-//      println(iter.next)
-//      println(iter.next)
-//      println(iter.hasNext)
-    }
   }
-
 }
